src/tools/border/requirements.txt
Pillow>=9.0.0
pika>=1.3.0
pydantic>=2.0.0